Fixed fetchApi bug that was returning code 403 when accesing any OL api; began prelims on d3.js chart
'node-fetch' was refusing to connect succesfully. After reviewing the code extensively, I decided to write a few test cases
in jest to compare 'node-fetch' and the native 'https' module. 'https' worked with my existing code, so I swapped it in for
'node-fetch'. The remaining code after this function call is now run asynchronously in the callback.